Добавил функционал тестов, прогресса, пофиксил прошлую ветку

+++ src/main/java/ru/sweetbun/BecomeAnyone/service/ModuleService.java
@@ -1,13 +1,12 @@
 package ru.sweetbun.BecomeAnyone.service;
 
+import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ru.sweetbun.BecomeAnyone.DTO.*;
 import ru.sweetbun.BecomeAnyone.entity.Course;
-import ru.sweetbun.BecomeAnyone.entity.Lesson;
 import ru.sweetbun.BecomeAnyone.entity.Module;
 import ru.sweetbun.BecomeAnyone.exception.ResourceNotFoundException;
 import ru.sweetbun.BecomeAnyone.repository.ModuleRepository;
@@ -19,44 +18,34 @@ import java.util.function.Function;
 import java.util.stream.Collectors;
 
 @Service
-@Transactional
+@RequiredArgsConstructor
 public class ModuleService {
 
     private final ModuleRepository moduleRepository;
-
+    @Lazy
     private final LessonService lessonService;
 
     private final ModelMapper modelMapper;
-
+    @Lazy
     private final CourseService courseService;
-
-    @Autowired
-    public ModuleService(ModuleRepository moduleRepository, @Lazy LessonService lessonService, ModelMapper modelMapper,
-                         @Lazy CourseService courseService) {
-        this.moduleRepository = moduleRepository;
-        this.lessonService = lessonService;
-        this.modelMapper = modelMapper;
-        this.courseService = courseService;
-    }
 
     public Module createModule(CreateModuleDTO moduleDTO, Long courseId) {
         Course course = courseService.getCourseById(courseId);
         return createModule(moduleDTO, course);
     }
 
+    @Transactional
     public void createModules(List<CreateModuleDTO> moduleDTOS, Course course) {
         for (CreateModuleDTO moduleDTO : moduleDTOS) {
             Module module = modelMapper.map(moduleDTO, Module.class);
             module.setCourse(course);
             course.getModules().add(module);
             moduleRepository.save(module);
-            List<CreateLessonDTO> lessonDTOS = moduleDTO.getLessons();
-            if (!lessonDTOS.isEmpty()) {
-                lessonService.createLessons(lessonDTOS, module);
-            }
+            lessonService.createLessons(moduleDTO.getLessons(), module);
         }
     }
 
+    @Transactional
     private Module createModule(CreateModuleDTO moduleDTO, Course course) {
         Module module = modelMapper.map(moduleDTO, Module.class);
         module.setCourse(course);
@@ -66,51 +55,55 @@ public class ModuleService {
 
     public Module getModuleById(Long id) {
         return moduleRepository.findById(id)
-                .orElseThrow(() -> new ResourceNotFoundException(Module.class.getSimpleName(), id));
+                .orElseThrow(() -> new ResourceNotFoundException(Module.class, id));
     }
 
     public List<Module> getAllModulesByCourse(Long courseId) {
         return moduleRepository.findAllByCourseOrderByOrderNumAsc(courseService.getCourseById(courseId));
     }
 
+    @Transactional
     public Module updateModule(UpdateModuleDTO updateModuleDTO, Long id) {
         Module module = getModuleById(id);
         modelMapper.map(updateModuleDTO, module);
         return moduleRepository.save(module);
     }
 
+    @Transactional
     public List<Module> updateModules(List<UpdateModuleInCourseDTO> moduleDTOS, Course course) {
         Map<Long, Module> currentModulesMap = course.getModules().stream()
                 .collect(Collectors.toMap(Module::getId, Function.identity()));
-        List<Module> updatedModules = new ArrayList<>();
 
-        for (UpdateModuleInCourseDTO moduleDTO : moduleDTOS) {
-            Long moduleDTOId = moduleDTO.getId();
-            if (moduleDTOId != null && currentModulesMap.containsKey(moduleDTOId)) {
-                Module module = currentModulesMap.get(moduleDTOId);
-                currentModulesMap.remove(moduleDTOId);
-                List<UpdateLessonInCourseDTO> lessonDTOS = moduleDTO.getLessons();
-                modelMapper.map(moduleDTO, module);
-                updatedModules.add(updateLessonsForModule(lessonDTOS, module));
-            } else {
-                List<UpdateLessonInCourseDTO> lessonDTOS = moduleDTO.getLessons();
-                Module newModule = modelMapper.map(moduleDTO, Module.class);
-                newModule.setCourse(course);
-                Module savedModule = moduleRepository.save(newModule);
-                updatedModules.add(updateLessonsForModule(lessonDTOS, savedModule));
-            }
-        }
-        moduleRepository.deleteAll(new ArrayList<>(currentModulesMap.values()));
+        List<Module> updatedModules = mergeModules(moduleDTOS, modelMapper, currentModulesMap, course,
+                moduleRepository, lessonService);
+        if (!currentModulesMap.isEmpty())
+            moduleRepository.deleteAll(new ArrayList<>(currentModulesMap.values()));
         return updatedModules;
     }
 
-    private Module updateLessonsForModule(List<UpdateLessonInCourseDTO> lessonDTOS, Module module) {
-        List<Lesson> lessons = lessonService.updateLessons(lessonDTOS, module);
-        module.setLessons(lessons);
-        return moduleRepository.save(module);
+    @Transactional
+    public static List<Module> mergeModules(List<UpdateModuleInCourseDTO> moduleDTOS, ModelMapper mapper,
+                                            Map<Long, Module> currentModulesMap, Course course,
+                                            ModuleRepository moduleRepository, LessonService lessonService) {
+        return moduleDTOS.stream().map(moduleDTO -> {
+            Module module;
+            Long moduleDTOId = moduleDTO.getId();
+
+            if (moduleDTOId != null && currentModulesMap.containsKey(moduleDTOId)) {
+                module = currentModulesMap.remove(moduleDTOId);
+                mapper.map(moduleDTO, module);
+            } else {
+                module = mapper.map(moduleDTO, Module.class);
+                module.setCourse(course);
+            }
+            module.setLessons(lessonService.updateLessons(moduleDTO.getLessons(), module));
+            return moduleRepository.save(module);
+        }).collect(Collectors.toList());
     }
 
-    public String deleteModuleById(Long id) {
+
+    @Transactional
+    public long deleteModuleById(Long id) {
         Module moduleToDelete = getModuleById(id);
         int orderNum = moduleToDelete.getOrderNum();
         moduleRepository.deleteById(id);
@@ -118,6 +111,6 @@ public class ModuleService {
         List<Module> modulesToUpdate = moduleRepository.findByOrderNumGreaterThan(orderNum);
         modulesToUpdate.forEach(module -> module.setOrderNum(module.getOrderNum() - 1));
         moduleRepository.saveAll(modulesToUpdate);
-        return "Module has been deleted with id: " + id;
+        return id;
     }
 }
